Update BusinessController.cs
Currently, the content within the module is identified by the core search, but clicking on the search results only leads to the module page and does not display the exact searched item. By adding this  line of the code, when the user clicks on a search result, the ID of the record that the result pertains to will be passed in the URL parameters.