Update List.ascx.cs for filtering record by DNN core Search
Currently, the content within the module is identified by the core search, but clicking on the search results only leads to the module page and does not display the exact searched item. By editing this file, when the user clicks on a search result, the exact searched record will be filtered in grid.